:tada: Extract data-generation to separate util
